CARDS-1422: Trigger email reminders 24 hours prior to appointment if survey hasn't been completed Make an abstract class with the common code in the two notification task
