+++ proms-resources/backend/src/main/java/io/uhndata/cards/proms/emailnotifications/InitialNotificationsTask.java
@@ -19,124 +19,25 @@
 
 package io.uhndata.cards.proms.emailnotifications;
 
-import java.util.Calendar;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-import org.apache.commons.lang3.StringUtils;
-import org.apache.sling.api.resource.LoginException;
-import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ResourceResolverFactory;
 import org.apache.sling.commons.messaging.mail.MailService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import io.uhndata.cards.auth.token.TokenManager;
-import io.uhndata.cards.emailnotifications.EmailUtils;
-import jakarta.mail.MessagingException;
 
-public class InitialNotificationsTask implements Runnable
+public class InitialNotificationsTask extends AbstractPromsNotification implements Runnable
 {
     private static final String PATIENT_NOTIFICATION_SUBJECT =
         "Welcome to DATAPRO: Answer your Pre-Appointment Questions";
 
-    /** Default log. */
-    private static final Logger LOGGER = LoggerFactory.getLogger(InitialNotificationsTask.class);
-
-    private static final String CARDS_HOST_AND_PORT =
-        StringUtils.defaultIfEmpty(System.getenv("CARDS_HOST_AND_PORT"), "localhost:8080");
-
-    private static final String CLINIC_SLING_PATH = "/Proms.html";
-
-    /** Provides access to resources. */
-    private final ResourceResolverFactory resolverFactory;
-
-    private final TokenManager tokenManager;
-
-    private final MailService mailService;
-
     InitialNotificationsTask(final ResourceResolverFactory resolverFactory,
         final TokenManager tokenManager, final MailService mailService)
     {
-        this.resolverFactory = resolverFactory;
-        this.tokenManager = tokenManager;
-        this.mailService = mailService;
+        super(resolverFactory, tokenManager, mailService);
     }
 
     @Override
     public void run()
     {
-        /*
-         * Query the `Visit information` Forms for appointments happening
-         * on day NOW+3. For example, if today is January 1st 2022, find
-         * all appointments on January 4th 2022. This task should run as
-         * close as possible to the start of the day, but after any
-         * upstream data is fetched (eg. from Data Lake), so that
-         * patients will have the maximum amount of time to complete
-         * their surveys.
-         */
-        final Calendar dateToQuery = Calendar.getInstance();
-        dateToQuery.add(Calendar.DATE, 3);
-        final Map<String, Object> parameters =
-            Collections.singletonMap(ResourceResolverFactory.SUBSERVICE, "EmailNotifications");
-        try (ResourceResolver resolver = this.resolverFactory.getServiceResourceResolver(parameters)) {
-            Iterator<Resource> appointmentResults = AppointmentUtils.getAppointmentsForDay(resolver, dateToQuery);
-            while (appointmentResults.hasNext()) {
-                Resource appointmentDate = appointmentResults.next();
-                Resource appointmentForm = AppointmentUtils.getFormForAnswer(resolver, appointmentDate);
-                if (appointmentForm == null) {
-                    continue;
-                }
-                // Get the Patient Subject associated with this appointment Form
-                Resource patientSubject = AppointmentUtils.getRelatedSubjectOfType(
-                    resolver, appointmentForm, "/SubjectTypes/Patient");
-                Resource visitSubject = AppointmentUtils.getRelatedSubjectOfType(
-                    resolver, appointmentForm, "/SubjectTypes/Patient/Visit");
-                String patientEmailAddress = AppointmentUtils.getPatientConsentedEmail(resolver, patientSubject);
-                if (patientEmailAddress == null) {
-                    continue;
-                }
-                String emailTextTemplate = AppointmentUtils.getVisitEmailTemplate(resolver, visitSubject, "72h.txt");
-                if (emailTextTemplate == null) {
-                    continue;
-                }
-                String patientFullName = AppointmentUtils.getPatientFullName(resolver, patientSubject);
-                Calendar tokenExpiryDate = AppointmentUtils.parseDate(appointmentDate.getValueMap().get("value", ""));
-                tokenExpiryDate.add(Calendar.DATE, 1);
-                atMidnight(tokenExpiryDate);
-                String surveysLink = "https://" + CARDS_HOST_AND_PORT + CLINIC_SLING_PATH + "?auth_token="
-                    + this.tokenManager.create(
-                        "patient",
-                        tokenExpiryDate,
-                        Collections.singletonMap(
-                            "cards:sessionSubject",
-                            visitSubject.getPath()
-                        )
-                    ).getToken();
-                // Send the Initial Notification Email
-                Map<String, String> valuesMap = new HashMap<>();
-                valuesMap.put("surveysLink", surveysLink);
-                String emailTextBody = EmailUtils.renderEmailTemplate(emailTextTemplate, valuesMap);
-                try {
-                    EmailUtils.sendNotificationEmail(this.mailService, patientEmailAddress,
-                        patientFullName, PATIENT_NOTIFICATION_SUBJECT, emailTextBody);
-                } catch (MessagingException e) {
-                    LOGGER.warn("Failed to send Initial Notification Email");
-                }
-            }
-        } catch (LoginException e) {
-            LOGGER.warn("Failed to results.next().getPath()");
-        }
-    }
-
-    private void atMidnight(final Calendar c)
-    {
-        c.set(Calendar.HOUR_OF_DAY, 0);
-        c.set(Calendar.MINUTE, 0);
-        c.set(Calendar.SECOND, 0);
-        c.set(Calendar.MILLISECOND, 0);
+        sendNotification(3, "72h.txt", PATIENT_NOTIFICATION_SUBJECT);
     }
 }
